GameMap 의 출력 포멧 관련 책임을 OutputView 로 위임
GameMap 은 사다리게임 맵 데이터만 저장하는데 집중할 수 있도록 변경하였습니다.
